Vectorize some functions
calculateLatestObjectVersion and getSpecificObjectVersion are now
both vectorized with respect to their 'keyList' argument so that
we don't have to repeatedly read the version file for a list of
keys.

This also implies that 'local.file.path' and 'local.file.path.SIG'
are also vectorized with respect to their 'key' arguments.

Also vectorized 'key2filename' so that a vector of keys can be
passed.