Simplify 'readVersionFileLine'; add 'where' argument to 'versionFile'
readVersionFileLine can be simplified by only reading in as many
lines as equal to the repository version.  Then, we always take
the last element of the character vector.

Giving 'versionFile' a 'where' argument allows us to reference
both the local and remote copies of the version file for 'remoteDB'
databases.